Got the "read in" and "find words" parts working
I got most of the code running (I should have done more commits... oops). The read in part of the code is pretty standard, though it is designed to handle arbitrary amouts of whitespace between the letters and m/n (to make it more robust).

The "find words" part of the code works as follows: it cycles through the words. For each word, it checks the rows, columns, desceding diagonals, and ascending diagonals for matches. The only tricky part of this was getting the diagional checks to work. I added lengthy comments to explain my thought process.

The code works, it finds all of the words. What's left is to write these results to a file (and clean up my code). Expect all of that in the next commit(s).