Made comment revision + minor tweaks
I updated and corrected many of the comments. I also made a few actual changes to the code. For one, I modified write_locations_to_file so that the width of the printed characters is determined dynamically at runtime rather than just using a fixed width of 3. See WordSearch.py:390-410 as well as 418-420

These revisions allowed me to better format found.txt (see the updated file).

This code runs and appears to give the correct results.